+++ frontend/src/app/pages/exercises/exercise/shared/partial-export/open-partial-export-selection-modal.ts
@@ -1,0 +1,8 @@
+import type { NgbModal } from '@ng-bootstrap/ng-bootstrap';
+import { PartialExportModalComponent } from './partial-export-modal/partial-export-modal.component';
+
+export function openPartialExportModal(ngbModalService: NgbModal) {
+    ngbModalService.open(PartialExportModalComponent, {
+        size: 'm',
+    });
+}