+++ frontend/src/app/pages/landing-page/landing-page/landing-page.component.ts
@@ -72,7 +72,6 @@ export class LandingPageComponent {
                 (type === 'complete'
                     ? StateExport
                     : PartialExport) as Constructor<
-                    // eslint-disable-next-line @typescript-eslint/no-unnecessary-type-arguments
                     PartialExport | StateExport
                 >,
                 importPlain
@@ -97,9 +96,12 @@ export class LandingPageComponent {
                     break;
                 }
                 case 'partial': {
-                    throw new Error(
-                        'Dieser Typ kann zur Zeit nicht importiert werden.'
-                    );
+                    this.messageService.postMessage({
+                        color: 'danger',
+                        title: 'Unerlaubter Importtyp',
+                        body: 'Dieser Typ kann nur innerhalb einer Übung importiert werden.',
+                    });
+                    break;
                 }
             }
         } catch (error: unknown) {