+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project does **not** adhere to [Semantic Versioning](https://semver.org
 - this software has now a license (with some exceptions), see LICENSE-README.md
     - the software includes a file with third party acknoledgments, a license.html and the docker container will include a git archive of the source code downloadable in zip format
     - new npm run commands for licensing (third party acknowledgements) and creating source code archive
+- Add possibility to export only templates from an exercise and import them in another one.
+    - Patients, vehicles, and image templates can be selected.
+    - Imports can either overwrite existing templates or be added to the exercise.
 - Vehicle templates can now be edited, added, and deleted from within an exercise.
 
 ### Changed